requirements.txt
colorama==0.4.6
pillow==12.1.1
tqdm==4.67.3
